Refactor backend structure for multiple commands (Resolves #40)
This separates out the core commands from the database/routing
structures to allow multiple commands. Testing should also be easier,
but some changes may need to be made there (such as abstracting the
creation/storage of the DB/Router instances).
