+++ backend/cmd/server/main.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"backend/src"
+)
+
+func main() {
+	//TODO: Error handling
+	src.InitDatabase()
+	src.InitRouter()
+	_ = src.Router.Run(":8080")
+}
